requirements.txt
telethon
requests
colorama
python-dotenv